+++ CHANGELOG.md
@@ -5,6 +5,17 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Changed
+
+- the `xml:base` specified on the `rdf:RDF` element is stored in the `base_uri`
+  field of the decoded `RDF.Graph` 
+
+[Compare v0.1.0...HEAD](https://github.com/rdf-elixir/rdf-xml-ex/compare/v0.1.0...HEAD)
+
+
+
 ## v0.1.0 - 2020-11-16
 
 Initial release